+++ apps/storybook/tailwind.config.ts
@@ -4,7 +4,7 @@ import type { Config } from "tailwindcss";
 const config: Config = {
   content: [
     "../../packages/components/*/src/*.ts",
-    "../../packages/components/*/src/*.tsx",
+    "../../packages/components/*/src/**/*.tsx",
     "../../packages/components/*/stories/**/*.tsx",
   ],
   plugins: [...jamsrui()],