+++ packages/components/editor/src/components/BlockEditor/BlockEditor.tsx
@@ -1,0 +1,71 @@
+"use client";
+
+import type { PureEditorContent } from "@tiptap/react";
+import { EditorContent } from "@tiptap/react";
+import { useMemo, useRef } from "react";
+
+import { LinkMenu } from "@/components/menus";
+
+import { useBlockEditor } from "@/hooks/useBlockEditor";
+
+import "@/styles/index.css";
+
+import { Sidebar } from "@/components/Sidebar";
+import { EditorContext } from "@/context/EditorContext";
+import ImageBlockMenu from "@/extensions/ImageBlock/components/ImageBlockMenu";
+import { ColumnsMenu } from "@/extensions/MultiColumn/menus";
+import { TableColumnMenu, TableRowMenu } from "@/extensions/Table/menus";
+import { ContentItemMenu } from "../menus/ContentItemMenu";
+import { TextMenu } from "../menus/TextMenu";
+import { EditorHeader } from "./components/EditorHeader";
+import type { TiptapProps } from "./types";
+
+export const BlockEditor = ({ content, onUpdate }: TiptapProps) => {
+  const menuContainerRef = useRef(null);
+  const editorRef = useRef<PureEditorContent | null>(null);
+  const { editor, characterCount, leftSidebar } = useBlockEditor({
+    content,
+    onUpdate,
+  });
+
+  const providerValue = useMemo(() => {
+    return {};
+  }, []);
+
+  if (!editor) {
+    return null;
+  }
+
+  return (
+    <EditorContext.Provider value={providerValue}>
+      <div className="flex h-full" ref={menuContainerRef}>
+        <Sidebar
+          isOpen={leftSidebar.isOpen}
+          onClose={leftSidebar.close}
+          editor={editor}
+        />
+        <div className="relative flex h-full flex-1 flex-col overflow-hidden">
+          <EditorHeader
+            characters={characterCount.characters()}
+            words={characterCount.words()}
+            isSidebarOpen={leftSidebar.isOpen}
+            toggleSidebar={leftSidebar.toggle}
+          />
+          <EditorContent
+            editor={editor}
+            // @ts-ignore
+            ref={editorRef}
+            className="flex-1 overflow-y-auto"
+          />
+          <ContentItemMenu editor={editor} />
+          <LinkMenu editor={editor} appendTo={menuContainerRef} />
+          <TextMenu editor={editor} />
+          <ColumnsMenu editor={editor} appendTo={menuContainerRef} />
+          <TableRowMenu editor={editor} appendTo={menuContainerRef} />
+          <TableColumnMenu editor={editor} appendTo={menuContainerRef} />
+          <ImageBlockMenu editor={editor} appendTo={menuContainerRef} />
+        </div>
+      </div>
+    </EditorContext.Provider>
+  );
+};