+++ packages/components/editor/tsup.config.ts
@@ -1,0 +1,14 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src"],
+  format: ["cjs", "esm"],
+  clean: true,
+  minify: true,
+  splitting: true,
+  esbuildOptions(options) {
+    options.banner = {
+      js: '"use client"',
+    };
+  },
+});